feat : quote Data 단 구현

+++ GLIM-AOS/app/src/main/java/com/ssafy/glim/core/domain/model/Glim.kt
@@ -4,7 +4,7 @@ data class Glim(
     val id: Long,
     val imgUrl: String,
     val isLike: Boolean,
-    val likes: Int,
+    val likes: Long,
     val bookTitle: String,
     val bookAuthor: String,
     val bookImgUrl: String,
